Refactor TypeHelpers for ObjC, Java
