+++ src/T4TemplateWriter/Templates/Java/Base/BaseTemplate.java.tt
