+++ src/T4TemplateWriter/Vipr.T4TemplateWriter.csproj
@@ -49,9 +49,8 @@
   <ItemGroup>
     <Compile Include="TemplateProcessor\CustomT4Host.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
-    <Compile Include="CodeHelpers\Java\PropertyHelper.cs" />
-    <Compile Include="CodeHelpers\Java\StringExtensions.cs" />
-    <Compile Include="CodeHelpers\ObjC\PropertyHelper.cs" />
+    <Compile Include="CodeHelpers\Java\TypeHelperJava.cs" />
+    <Compile Include="CodeHelpers\ObjC\TypeHelperObjC.cs" />
     <Compile Include="Extensions\OdcmModelExtensions.cs" />
     <Compile Include="PathWriters\PathWriterBase.cs" />
     <Compile Include="PathWriters\IPathWriter.cs" />
@@ -72,7 +71,7 @@
     <None Include=".config\TemplateWriterSettings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-    <None Include="Templates\Java\Base\BaseTemplateCustomHost.java.tt">
+    <None Include="Templates\Java\Base\BaseTemplate.java.tt">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Include="Templates\Java\Model\ComplexType.java.tt">
